+++ topic/src/main/java/tech/ydb/topic/impl/GrpcTopicRpc.java
@@ -7,12 +7,12 @@ import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 import javax.annotation.WillNotClose;
 
-import tech.ydb.core.Operations;
 import tech.ydb.core.Result;
 import tech.ydb.core.Status;
 import tech.ydb.core.grpc.GrpcReadWriteStream;
 import tech.ydb.core.grpc.GrpcRequestSettings;
 import tech.ydb.core.grpc.GrpcTransport;
+import tech.ydb.core.impl.operation.OperationManager;
 import tech.ydb.topic.TopicRpc;
 import tech.ydb.topic.YdbTopic;
 import tech.ydb.topic.v1.TopicServiceGrpc;
@@ -20,14 +20,17 @@ import tech.ydb.topic.v1.TopicServiceGrpc;
 
 /**
  * @author Nikolay Perfilov
+ * @author Kirill Kurdyukov
  */
 @ParametersAreNonnullByDefault
 public final class GrpcTopicRpc implements TopicRpc {
 
     private final GrpcTransport transport;
+    private final OperationManager operationManager;
 
     private GrpcTopicRpc(GrpcTransport transport) {
         this.transport = transport;
+        this.operationManager = transport.getOperationManager();
     }
 
     @Nullable
@@ -39,14 +42,14 @@ public final class GrpcTopicRpc implements TopicRpc {
     public CompletableFuture<Status> createTopic(YdbTopic.CreateTopicRequest request, GrpcRequestSettings settings) {
         return transport
                 .unaryCall(TopicServiceGrpc.getCreateTopicMethod(), settings, request)
-                .thenApply(Operations.statusUnwrapper(YdbTopic.CreateTopicResponse::getOperation));
+                .thenCompose(operationManager.statusUnwrapper(YdbTopic.CreateTopicResponse::getOperation));
     }
 
     @Override
     public CompletableFuture<Status> alterTopic(YdbTopic.AlterTopicRequest request, GrpcRequestSettings settings) {
         return transport
                 .unaryCall(TopicServiceGrpc.getAlterTopicMethod(), settings, request)
-                .thenApply(Operations.statusUnwrapper(YdbTopic.AlterTopicResponse::getOperation));
+                .thenCompose(operationManager.statusUnwrapper(YdbTopic.AlterTopicResponse::getOperation));
     }
 
     @Override
@@ -54,22 +57,26 @@ public final class GrpcTopicRpc implements TopicRpc {
                                                                                  GrpcRequestSettings settings) {
         return transport
                 .unaryCall(TopicServiceGrpc.getDescribeTopicMethod(), settings, request)
-                .thenApply(Operations.resultUnwrapper(YdbTopic.DescribeTopicResponse::getOperation,
-                        YdbTopic.DescribeTopicResult.class));
+                .thenCompose(
+                        operationManager.resultUnwrapper(
+                                YdbTopic.DescribeTopicResponse::getOperation,
+                                YdbTopic.DescribeTopicResult.class
+                        )
+                );
     }
 
     @Override
     public CompletableFuture<Status> dropTopic(YdbTopic.DropTopicRequest request, GrpcRequestSettings settings) {
         return transport
                 .unaryCall(TopicServiceGrpc.getDropTopicMethod(), settings, request)
-                .thenApply(Operations.statusUnwrapper(YdbTopic.DropTopicResponse::getOperation));
+                .thenCompose(operationManager.statusUnwrapper(YdbTopic.DropTopicResponse::getOperation));
     }
 
     @Override
     public GrpcReadWriteStream<
-        YdbTopic.StreamWriteMessage.FromServer,
-        YdbTopic.StreamWriteMessage.FromClient
-        > writeSession() {
+            YdbTopic.StreamWriteMessage.FromServer,
+            YdbTopic.StreamWriteMessage.FromClient
+            > writeSession() {
         return transport.readWriteStreamCall(TopicServiceGrpc.getStreamWriteMethod(),
                 GrpcRequestSettings.newBuilder().build());
     }